test: FeatureServiceEndpoint test update
